# docs/contracts/assessment.openapi.yaml
openapi: 3.1.0
info:
  title: ViaVitae Digital Assessment Funnel API
  version: 1.0.0
  description: |
    Contract for the "Request a Digital Assessment" lead funnel.

    SINGLE SOURCE OF TRUTH (ADR-WEB-005): viavitae-web generates TypeScript types
    from this file (`pnpm generate:types` -> src/generated/assessment.ts);
    viavitae-api generates Pydantic models from the same file. Never hand-edit the
    generated artefacts. A PR touching this file requires @IterVitae (four-eyes)
    review, a version bump, and regeneration by every consumer.

    GDPR (DPIA-001): collects BUSINESS contact data in a church vertical. Data
    minimisation applies — do NOT request data revealing the religious beliefs of
    an individual respondent. `organization_type` is a business attribute of the
    requesting organisation, never a personal special-category datum (Art. 9).

    FIX vs. the v1.1 draft (review 2026-09-06):
      1. Path is `/assessment`, not `/v1/assessment`. The server URL already ends
         in `/v1`; the draft produced `/v1/v1/assessment`.
      2. `consent_privacy` keeps `const: true`, so a `false` value is a SCHEMA
         violation and maps to 400 (not 422). 422 is reserved for payloads that
         are schema-valid but break a business rule.
      3. Idempotency clarified: replaying the SAME key+payload returns 202 with
         status `duplicate_ignored`; reusing a key with a DIFFERENT payload
         returns 409.
      4. 429 documents a `Retry-After` header.
      5. Added required `consent_notice_version` so the controller can store the
         exact privacy-notice text version the submitter agreed to, as mandated by
         DPIA-001 control R7 (demonstrable consent). viavitae-api is not yet
         implemented, so adding a required field is non-breaking.
servers:
  - url: https://api.viavitae.com/v1
    description: production
  - url: https://api-staging.viavitae.com/v1
    description: staging
tags:
  - name: assessment
paths:
  /assessment:
    post:
      operationId: submitAssessment
      summary: Submit a Digital Assessment request
      tags: [assessment]
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          description: |
            Client-generated UUID. Retries of the same submission MUST reuse the
            same key so the API can de-duplicate.
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssessmentRequest'
      responses:
        '202':
          description: |
            Accepted — lead created, or already existed for this Idempotency-Key
            (status `duplicate_ignored`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessmentAccepted'
        '400':
          description: |
            Validation error, including a `consent_privacy` value that is not
            `true` (schema `const` violation) and a honeypot that is non-empty
            when strict validation is enabled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Idempotency-Key reused with a different request payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: |
            Schema-valid but a business rule was violated (for example
            `organization_type: other_organization` with no `message`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limited.
          headers:
            Retry-After:
              description: Seconds to wait before retrying.
              schema:
                type: integer
                minimum: 1
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    OrganizationType:
      type: string
      description: |
        Type of ORGANIZATION requesting an assessment (business attribute, not a
        personal special-category datum). Field names and UI copy must never imply
        the personal religious affiliation of the respondent.
      enum:
        - basilica
        - cathedral
        - diocese
        - deanery
        - parish_church
        - funeral_services
        - cemetery_services
        - online_store
        - marketplace_vendor
        - other_organization
    CurrentSiteStatus:
      type: string
      enum: [no_website, outdated, needs_redesign, unhappy_with_provider]
    AssessmentGoal:
      type: string
      enum:
        - new_website
        - ecommerce
        - donations
        - crm_integration
        - ai_assistant
        - cemetery_map
        - maintenance_only
    AssessmentRequest:
      type: object
      additionalProperties: false
      required:
        - organization_type
        - contact_email
        - consent_privacy
        - consent_notice_version
        - locale
      properties:
        organization_type:
          $ref: '#/components/schemas/OrganizationType'
        organization_name:
          type: string
          minLength: 2
          maxLength: 200
        current_site_url:
          type: [string, 'null']
          format: uri
          maxLength: 2048
        current_status:
          $ref: '#/components/schemas/CurrentSiteStatus'
        goals:
          type: array
          maxItems: 5
          items:
            $ref: '#/components/schemas/AssessmentGoal'
        message:
          type: [string, 'null']
          maxLength: 2000
        contact_name:
          type: string
          minLength: 2
          maxLength: 120
        contact_email:
          type: string
          format: email
          maxLength: 254
        contact_phone:
          type: [string, 'null']
          pattern: '^\+?[0-9 ()-]{6,20}$'
          maxLength: 20
        preferred_language:
          type: string
          description: Preferred CONTACT language (not the website locale).
          enum: [lt, en, ru, pl, de]
        locale:
          type: string
          description: UI locale of the submitting form (analytics only).
          enum: [lt, en, ru]
        utm:
          type: [object, 'null']
          additionalProperties: false
          properties:
            source: { type: string, maxLength: 120 }
            medium: { type: string, maxLength: 120 }
            campaign: { type: string, maxLength: 120 }
        consent_privacy:
          type: boolean
          const: true
          description: Must be true. A false value is a 400 schema violation.
        consent_notice_version:
          type: string
          minLength: 1
          maxLength: 40
          description: |
            Identifier of the exact privacy-notice text version shown to and
            accepted by the submitter (e.g. "assessment-privacy-2026-09"). Stored
            with the consent record to make consent demonstrable — DPIA-001 R7.
            Supplied by the web tier from lib/constants.ts CONSENT_NOTICE_VERSION.
        consent_marketing:
          type: boolean
          default: false
        honeypot:
          type: string
          maxLength: 0
          description: |
            Must be empty. A non-empty value indicates a bot; the edge proxy
            (app/api/assessment) short-circuits with a synthetic 202 and never
            forwards it to the API.
    AssessmentAccepted:
      type: object
      additionalProperties: false
      required: [assessment_id, status]
      properties:
        assessment_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [accepted, duplicate_ignored]
        next_steps:
          type: object
          additionalProperties: false
          properties:
            confirmation_email_sent:
              type: boolean
            booking_url:
              type: [string, 'null']
              format: uri
              description: Cal.com scheduling link, present only when consent_privacy is true.
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            fields:
              type: object
              additionalProperties:
                type: string